Adding tests
Apply fixes from StyleCI

more tests

and more tests

Apply fixes from StyleCI

fix command

update travis script

removing server provider adding the new serve command to foundation

Apply fixes from StyleCI

phpstan fixes

adding first test for serve

Apply fixes from StyleCI

fix test
