+++ src/Viserio/Bridge/Twig/composer.json
@@ -19,7 +19,6 @@
     ],
     "require": {
         "php"                                : "^7.1",
-        "symfony/polyfill-mbstring"          : "^1.0",
         "twig/twig"                          : "~2.4"
     },
     "require-dev": {
