+++ src/Viserio/Component/Foundation/Console/Resources/router.php
