+++ src/Viserio/Component/Foundation/Provider/ConsoleCommandsServiceProvider.php
@@ -4,9 +4,11 @@ namespace Viserio\Component\Foundation\Provider;
 
 use Interop\Container\ServiceProvider;
 use Psr\Container\ContainerInterface;
+use Symfony\Component\Process\Process;
 use Viserio\Component\Console\Application;
 use Viserio\Component\Foundation\Console\Command\DownCommand;
 use Viserio\Component\Foundation\Console\Command\KeyGenerateCommand;
+use Viserio\Component\Foundation\Console\Command\ServeCommand;
 use Viserio\Component\Foundation\Console\Command\UpCommand;
 
 class ConsoleCommandsServiceProvider implements ServiceProvider
@@ -39,6 +41,10 @@ class ConsoleCommandsServiceProvider implements ServiceProvider
                 new UpCommand(),
                 new KeyGenerateCommand(),
             ]);
+
+            if (class_exists(Process::class)) {
+                $console->add(new ServeCommand());
+            }
         }
 
         return $console;
