--- src/Viserio/Provider/WebServer/README.md
@@ -1,36 +1,0 @@
-<h1 align="center">Viserio Twig Provider</h1>
-<p align="center">
-    <a href="https://github.com/narrowspark/framework/releases"><img src="https://img.shields.io/packagist/v/narrowspark/framework.svg?style=flat-square"></a>
-    <a href="https://php.net/"><img src="https://img.shields.io/badge/php-%5E7.1.0-8892BF.svg?style=flat-square"></a>
-    <a href="https://codecov.io/gh/narrowspark/framework"><img src="https://img.shields.io/codecov/c/github/narrowspark/framework/master.svg?style=flat-square"></a>
-    <a href="https://gitter.im/narrowspark/framework"><img src="https://img.shields.io/gitter/room/nwjs/nw.js.svg?style=flat-square"></a>
-    <a href="http://opensource.org/licenses/MIT"><img src="https://img.shields.io/badge/license-MIT-brightgreen.svg?style=flat-square"></a>
-</p>
-
-
-
-> **Note:** This package is part of the [Narrowspark framework](http://github.com/narrowspark/framework). <br> If you want to build an application using Narrowspark, visit the main [source](https://github.com/narrowspark/framework).
-
-Installation
--------------
-
-Use [Composer](https://getcomposer.org/) to install this package:
-
-```sh
-composer require viserio/twig-provider
-```
-
-Official Documentation
--------------
-
-Documentation for the framework can be found on the [Narrowspark website](http://narrowspark.com/docs).
-
-Contributing
--------------
-Issues for this package shall be posted on [Narrowspark framework issues](http://github.com/narrowspark/framework/issues). <br>
-Thank you for considering contributing to the Narrowspark framework! The contribution guide can be found in the [Narrowspark contributing](/CONTRIBUTING.md).
-
-License
--------------
-
-The Narrowspark framework is open-sourced software licensed under the [MIT license](http://opensource.org/licenses/MIT).
