--- src/Viserio/Provider/WebServer/composer.json
@@ -1,53 +1,0 @@
-{
-    "name"       : "viserio/web-server-provider",
-    "type"       : "viserio-provider",
-    "description": "The Viserio WebServer provider.",
-    "keywords"   : ["narrowspark", "viserio", "webserver", "web server", "provider", "php server"],
-    "license"    : "MIT",
-    "homepage"   : "http://github.com/narrowspark/framework",
-    "support"    : {
-        "issues": "https://github.com/narrowspark/framework/issues",
-        "source": "https://github.com/narrowspark/framework"
-    },
-    "authors"    : [
-        {
-          "name"    : "Daniel Bannert",
-          "email"   : "d.bannert@anolilab.de",
-          "homepage": "http://www.anolilab.de",
-          "role"    : "Developer"
-        }
-    ],
-    "require": {
-        "php"                                       : "^7.1",
-        "symfony/process"                           : "~3.3",
-        "viserio/console"                           : "self.version",
-        "viserio/options-resolver"                  : "self.version"
-    },
-    "require-dev": {
-        "container-interop/service-provider"        : "^0.3",
-        "mockery/mockery"                           : "^0.9",
-        "narrowspark/testing-helper"                : "^3.0",
-        "phpunit/phpunit"                           : "^6.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "Viserio\\Provider\\WebServer\\"        : ""
-        },
-        "exclude-from-classmap"                     : ["/Tests/"]
-    },
-    "autoload-dev": {
-        "psr-4": {
-            "Viserio\\Provider\\WebServer\\Tests\\" : "Tests/"
-        }
-    },
-    "suggest": {
-        "container-interop/service-provider"        : "Required to use service-provider (^0.3)."
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master"                            : "1.0-dev"
-        }
-    },
-    "minimum-stability"                             : "dev",
-    "prefer-stable"                                 : true
-}
